Move cx to utils
It belongs in utils instead of theme.
