+++ src/theme/form/FormControl.tsx
@@ -1,5 +1,5 @@
 import { h, FunctionComponent } from "preact";
-import { cx } from "../cx";
+import { cx } from "../../util/cx";
 
 export const FormControl: FunctionComponent = ({ children }) => (
     <div className={cx("w-full")}>{children}</div>
