+++ src/theme/form/FormLabel.tsx
@@ -1,5 +1,5 @@
 import { h, FunctionComponent } from "preact";
-import { cx } from "../cx";
+import { cx } from "../../util/cx";
 import { labeled } from "../styles";
 
 export const FormLabel: FunctionComponent<{ id: string }> = ({
