+++ src/theme/form/HelpText.tsx
@@ -1,5 +1,5 @@
 import { FunctionComponent, h } from "preact";
-import { cx } from "../cx";
+import { cx } from "../../util/cx";
 
 export const HelpText: FunctionComponent = ({ children }) => (
     <p className={cx("text-gray-600", "leading-6")}>{children}</p>
