+++ src/theme/toolbar/Toolbar.tsx
@@ -1,5 +1,5 @@
 import { FunctionComponent, h } from "preact";
-import { cx } from "../cx";
+import { cx } from "../../util/cx";
 
 export const Toolbar: FunctionComponent = ({ children }) => (
     <nav className={cx("fixed", "top-0", "left-0", "p-2", "space-y-2")}>
